+++ kubejs/server_scripts/Greg Recipies/New Materials/casings.js
@@ -33,7 +33,8 @@ ServerEvents.recipes(event => {
     casingass('copper_casing_wood', '6x gtceu:copper_plate', '#forge:stripped_wood', '2x create:copper_casing')
     casingass('steel_casing_log', '6x gtceu:steel_plate', '#forge:stripped_logs', '2x tfmg:steel_casing')
     casingass('steel_casing_wood', '6x gtceu:steel_plate', '#forge:stripped_wood', '2x tfmg:steel_casing')
-    casingass('kubejs:agri_casing', 'gtceu:al_6xn_plate', 'gtceu:al_6xn_frame', '2x kubejs:agri_casing')
+    casingass('kubejs:agri_casing', '6x gtceu:al_6xn_plate', 'gtceu:al_6xn_frame', '2x kubejs:agri_casing')
+    casingass('gtceu:atomic_casing', '6x gtceu:extreme_plate', 'gtceu:extreme_frame', '2x gtceu:atomic_casing')
     //casingass('electrum_casing_log', '6x gtceu:electrum_plate', 'kubejs:stripped_treated_wood_log', '2x kubejs:electrum_casing')
     //casingass('electrum_casing_wood', '6x gtceu:electrum_plate', 'kubejs:stripped_treated_wood', '2x kubejs:electrum_casing')
 
@@ -46,6 +47,7 @@ ServerEvents.recipes(event => {
     casingcraft('tfmg:steel_casing', 'gtceu:steel_plate', '#forge:stripped_logs')
     casingcraft('tfmg:steel_casing', 'gtceu:steel_plate', '#forge:stripped_wood')
     casingcraft('kubejs:agri_casing', 'gtceu:al_6xn_plate', 'gtceu:al_6xn_frame')
+    casingcraft('gtceu:atomic_casing', 'gtceu:extreme_plate', 'gtceu:extreme_frame')
     //casingcraft('kubejs:electrum_casing', 'gtceu:electrum_plate', 'kubejs:stripped_treated_wood_log')
     //casingcraft('kubejs:electrum_casing', 'gtceu:electrum_plate', 'kubejs:stripped_treated_wood')
 })