+++ changelog.md
@@ -10,11 +10,15 @@
   * Fixed missing lang for Soul Soil ores
 * Tags
   * Fixed missing tags for Soul Soil ores, now correctly minable with pickaxe or shovel
+* Recipes
+  * Removed the ability to use non gregified piston recipe
 * Machines
   * Singleblock Machines
     * Added Decay Hastener
     * Added Fuel Reprocessor
     * Added Isotope Separator
+  * Multiblock Machines
+    * Added Large Decay Hastener
 * Materials
   * Additions
     * Added Berkelium
@@ -34,7 +38,6 @@
     * Added Dimensional Blend
     * Added End Steel
     * Added Energetic Alloy
-    * Added Energetic Blend
     * Added Extreme Alloy
     * Added Ferroboron
     * Added Fluix
@@ -43,6 +46,7 @@
     * Added Hard Carbon
     * Added Lithium Manganese Dioxide
     * Added Manganosite
+    * Added Nepheline
     * Added Neptunium
     * Added Niobium-Tin
     * Added Polonium
@@ -52,6 +56,7 @@
     * Added Radium
     * Added Rhodochrosite
     * Added Soularium
+    * Added Titanite
     * Added Thallium
     * Added Thermoconducting Alloy
     * Added Tough Alloy
@@ -66,6 +71,19 @@
     * Renamed Enriched Certus Quartz to Charged Certus Quartz
   * Modifications
     * Added material liquid for all materials with an ingot form
+    * Desh now requires an Mv EBF with Cupronickel coils
+    * Ostrum now requires an Hv EBF with Kanthal Coils
+    * Calorite now requires an Ev EBF with Nichrome Coils
+    * Moon Deepslate no longer contains quicklime, can now be processed in an electrolyzer
+    * Added ore to Biotite
+* Ore Veins
+  * Additions
+    * Added Carobbiite (Carobbiite) vein to Mars
+    * Added Nepheline (Nepheline, Biotite, Titanite, Villiaumite) vein to Mars
+    * Added Fluorspar (Fluorspar, Barite, Calcite) vein to Mars
+    * Added Rhodochrosite (Rhodochrosite, Silver) vein to Mars
+  * Modifications
+    * Fixed Naquadah, Sheldonite, Pitchblende, Scheelite veins not spawning on Mars
 * Resource Packs
   * Additions
     * Added Better GUI for Sophisticated Backpacks
@@ -80,5 +98,6 @@
   * Mod Updates
     * Amendments v1.2.11
     * GregTechCEu Modern v1.4.1+patch.2
+    * Just Enough Items v15.19.0.88
     * LDLib v1.0.27.b
     * Supplementaries v2.8.17